feat(#2018): action first, and the run card split by its own seams
Founder's ask: "move the action first where they see what they have to
do … show start only and then reveal each section".

`production-run-card.tsx` was 1,698 lines — the largest file in the app —
and everything the partner must DO sat at the bottom of nine stacked
sections inside it. Split into nine files, none over ~600 lines:

  lib/run-formatting.ts        formatters (pure)
  lib/run-phase.ts             phase + next action (pure, 12 tests)
  run-progress-stepper.tsx     STEPS + ProgressStepper
  run-timeline-item.tsx        TimelineItem
  run-info-banner.tsx          InfoBanner
  run-next-action.tsx          NEW — the pinned action block
  finish-run-form.tsx          FinishRunForm
  complete-run-form.tsx        CompleteRunForm
  inline-task-card.tsx         InlineTaskCard + InlineSubtaskRow

`RunNextAction` replaces TWO independent derivations of one state: the
header's `primaryAction` chain (the button) and `getStageGuidance` (a
banner saying the same thing in words). The label and the sentence under
it now come from one call, so they cannot drift. The guidance module is
deleted rather than left unreferenced.

🔴 The action does NOT derive from the phase, and the difference is the
point. The card's `can*` guards key on `status === "in_progress"`, so a
run with `accepted_at` set whose status is still `sent_to_partner` is in
phase "accepted" while `canStart` refuses it. Deriving the button from
the phase would offer Start for a run the API rejects. Phase decides what
the page REVEALS; `getRunNextAction` decides what the partner may DO, and
mirrors the guards exactly — including `actionable: false`, the
defence-in-depth guard for a cancelled design assignment. Both are tested.

i18n from the start, per #2022 — 20 new keys under `partner.workOrders`,
and the key types are LITERAL unions rather than `string`, so `t()` checks
them against the generated catalogue. A missing key now fails the build
instead of rendering its own name at the top of the partner's page.

Verify: npx vitest --run src/lib/run-phase.test.ts (12)
