+++ apps/partner-ui/src/components/work-orders/production-run-card.tsx
@@ -1,290 +1,30 @@
 import {
   Alert,
   Badge,
-  Button,
-  Checkbox,
   Container,
-  Drawer,
-  FocusModal,
   Heading,
-  Input,
-  Select,
   StatusBadge,
   Text,
-  Textarea,
-  Tooltip,
   clx,
   toast,
   usePrompt,
 } from "@medusajs/ui"
-import { InformationCircleSolid, ExclamationCircle } from "@medusajs/icons"
 import { useState } from "react"
-import { Link } from "react-router-dom"
-
-import { resolveRunMaterialOptions } from "../../lib/run-materials"
-import { planCompletionOutput } from "../../lib/completion-output"
 import { PartnerDesign } from "../../hooks/api/partner-designs"
-import {
-  useAcceptPartnerProductionRun,
-  useStartPartnerProductionRun,
-  useFinishPartnerProductionRun,
-  useCompletePartnerProductionRun,
-} from "../../hooks/api/partner-production-runs"
-import {
-  useAcceptPartnerAssignedTask,
-  useFinishPartnerAssignedTask,
-  useCompletePartnerAssignedTaskSubtask,
-} from "../../hooks/api/partner-assigned-tasks"
+import { useAcceptPartnerProductionRun, useCompletePartnerProductionRun, useFinishPartnerProductionRun, useStartPartnerProductionRun } from "../../hooks/api/partner-production-runs"
 import { useMe } from "../../hooks/api/users"
-import { getStatusBadgeColor } from "../../lib/status-badge"
-import { getReassignmentNotice } from "../../lib/reassignment-notice"
 import { extractErrorMessage } from "../../lib/extract-error-message"
+import { getReassignmentNotice } from "../../lib/reassignment-notice"
 import { GoodsTransferSection } from "./goods-transfer-section"
-
-// ── Helpers ──────────────────────────────────────────────────────────
-
-/** Format a cost value with currency code */
-function formatCost(value: number | null | undefined, currency?: string | null): string {
-  if (value == null) return "-"
-  const code = (currency || "").toUpperCase()
-  return code ? `${code} ${value}` : String(value)
-}
-
-function formatDateTime(dateStr: string | null | undefined): string {
-  if (!dateStr) return "-"
-  return new Date(dateStr).toLocaleString("en-US", {
-    month: "short",
-    day: "numeric",
-    hour: "2-digit",
-    minute: "2-digit",
-  })
-}
-
-function formatFullDate(dateStr: string | null | undefined): string {
-  if (!dateStr) return ""
-  return new Date(dateStr).toLocaleString("en-US", {
-    weekday: "long",
-    year: "numeric",
-    month: "long",
-    day: "numeric",
-    hour: "2-digit",
-    minute: "2-digit",
-  })
-}
-
-function durationBetween(from: string | null | undefined, to: string | null | undefined): string {
-  if (!from || !to) return ""
-  const diffMs = new Date(to).getTime() - new Date(from).getTime()
-  if (diffMs < 0) return ""
-  const mins = Math.floor(diffMs / 60000)
-  if (mins < 60) return `${mins}m`
-  const hrs = Math.floor(mins / 60)
-  if (hrs < 24) return `${hrs}h ${mins % 60}m`
-  const days = Math.floor(hrs / 24)
-  return `${days}d ${hrs % 24}h`
-}
-
-function getTargetDateStatus(targetDate: string | null | undefined): {
-  label: string
-  color: "red" | "orange" | "grey"
-} | null {
-  if (!targetDate) return null
-  const target = new Date(targetDate)
-  const now = new Date()
-  const diffDays = Math.ceil((target.getTime() - now.getTime()) / (1000 * 60 * 60 * 24))
-  const formatted = target.toLocaleDateString("en-US", { month: "short", day: "numeric" })
-
-  if (diffDays < 0) return { label: `Overdue (${formatted})`, color: "red" }
-  if (diffDays <= 2) return { label: `Due ${formatted}`, color: "orange" }
-  return { label: `Due ${formatted}`, color: "grey" }
-}
-
-/** Map run status to StatusBadge color */
-const runStatusColor = (status: string): "green" | "orange" | "red" | "blue" | "grey" => {
-  switch (status) {
-    case "completed": return "green"
-    case "in_progress": return "orange"
-    case "sent_to_partner": return "blue"
-    case "cancelled": return "red"
-    default: return "grey"
-  }
-}
-
-/** Returns contextual guidance for the current stage */
-function getStageGuidance(
-  run: any,
-  isSample: boolean
-): { title: string; description: string } | null {
-  const status = String(run.status || "")
-  if (status === "cancelled" || status === "completed") return null
-
-  if (status === "sent_to_partner") {
-    return {
-      title: "Accept this run to get started",
-      description: "Review the details and confirm you'll handle this work.",
-    }
-  }
-  if (status === "in_progress" && !run.started_at) {
-    return {
-      title: "Mark as started when you begin working",
-      description: "This helps track production timelines accurately.",
-    }
-  }
-  if (status === "in_progress" && run.started_at && !run.finished_at) {
-    if (isSample) {
-      return {
-        title: "Log materials as you work",
-        description:
-          "Track all materials used — this data will be used to calculate the design's cost estimate. Mark finished when done.",
-      }
-    }
-    return {
-      title: "Mark finished when work is done",
-      description: "Once finished, the admin team will review your work before final completion.",
-    }
-  }
-  if (status === "in_progress" && run.finished_at) {
-    if (isSample) {
-      return {
-        title: "Complete with cost details",
-        description:
-          "Log final material usage and your cost estimate. This is critical for pricing — be thorough.",
-      }
-    }
-    return {
-      title: "Complete the run",
-      description: "Log any remaining material usage and your production cost to finalize.",
-    }
-  }
-  return null
-}
-
-// ── Progress Steps ──────────────────────────────────────────────────
-
-const STEPS = [
-  { key: "received", label: "Received" },
-  { key: "accepted", label: "Accepted" },
-  { key: "started", label: "Started" },
-  { key: "finished", label: "Finished" },
-  { key: "completed", label: "Completed" },
-]
-
-const ProgressStepper = ({ run }: { run: any }) => {
-  const status = String(run.status || "")
-  if (status === "cancelled") return null
-
-  // Determine step from timestamps first, fall back to status field
-  // (handles cases where status was set via task-completion subscriber without lifecycle timestamps)
-  let currentIdx = 0
-  if (run.completed_at || status === "completed") currentIdx = 4
-  else if (run.finished_at) currentIdx = 3
-  else if (run.started_at || status === "in_progress") currentIdx = 2
-  else if (run.accepted_at) currentIdx = 1
-  else if (status === "sent_to_partner") currentIdx = 0
-
-  return (
-    <div className="flex items-center gap-1 px-6 py-3">
-      {STEPS.map((step, idx) => {
-        const isDone = idx <= currentIdx
-        const isCurrent = idx === currentIdx
-        return (
-          <div key={step.key} className="flex items-center gap-1 flex-1">
-            <div className="flex flex-col items-center flex-1">
-              <div
-                className={clx("h-1.5 w-full rounded-full", {
-                  "bg-ui-fg-interactive": isDone,
-                  "bg-ui-border-base": !isDone,
-                })}
-              />
-              <Text
-                size="xsmall"
-                className={clx("mt-1", {
-                  "text-ui-fg-base font-medium": isCurrent,
-                  "text-ui-fg-subtle": isDone && !isCurrent,
-                  "text-ui-fg-muted": !isDone,
-                })}
-              >
-                {step.label}
-              </Text>
-            </div>
-          </div>
-        )
-      })}
-    </div>
-  )
-}
-
-// ── Timeline Item (dot + line pattern) ──────────────────────────────
-
-const TimelineItem = ({
-  label,
-  dateStr,
-  duration,
-  isLast,
-  dotColor = "bg-ui-tag-neutral-icon",
-}: {
-  label: string
-  dateStr: string
-  duration?: string
-  isLast?: boolean
-  dotColor?: string
-}) => (
-  <div className="grid grid-cols-[20px_1fr] items-start gap-2">
-    <div className="flex size-full flex-col items-center gap-y-0.5">
-      <div className="flex size-5 items-center justify-center">
-        <div className="bg-ui-bg-base shadow-borders-base flex size-2.5 items-center justify-center rounded-full">
-          <div className={clx("size-1.5 rounded-full", dotColor)} />
-        </div>
-      </div>
-      {!isLast && <div className="bg-ui-border-base w-px flex-1" />}
-    </div>
-    <div className={clx({ "pb-3": !isLast })}>
-      <div className="flex items-center justify-between gap-2">
-        <Text size="small" leading="compact" weight="plus">
-          {label}
-        </Text>
-        <Tooltip content={formatFullDate(dateStr)}>
-          <Text size="small" leading="compact" className="text-ui-fg-subtle text-right cursor-default">
-            {formatDateTime(dateStr)}
-          </Text>
-        </Tooltip>
-      </div>
-      {duration && (
-        <Text size="xsmall" className="text-ui-fg-muted">{duration}</Text>
-      )}
-    </div>
-  </div>
-)
-
-// ── Info Banner (Medusa pattern) ────────────────────────────────────
-
-const InfoBanner = ({
-  title,
-  description,
-  variant = "info",
-}: {
-  title: string
-  description: string
-  variant?: "info" | "warning"
-}) => (
-  <div className="flex items-start gap-x-3 rounded-xl border border-ui-border-base bg-ui-bg-subtle px-4 py-3 mx-6 my-3">
-    {variant === "warning" ? (
-      <ExclamationCircle className="mt-0.5 shrink-0 text-ui-tag-orange-icon" />
-    ) : (
-      <InformationCircleSolid className="mt-0.5 shrink-0 text-ui-fg-interactive" />
-    )}
-    <div className="flex flex-col gap-y-0.5">
-      <Text size="small" weight="plus" className="text-ui-fg-base">
-        {title}
-      </Text>
-      <Text size="xsmall" className="text-ui-fg-subtle">
-        {description}
-      </Text>
-    </div>
-  </div>
-)
-
-// ── Production Run Card ─────────────────────────────────────────────
+import { RunNextAction } from "./run-next-action"
+import type { RunActionKey } from "../../lib/run-phase"
+import { durationBetween, formatCost, getTargetDateStatus, runStatusColor } from "../../lib/run-formatting"
+import { ProgressStepper } from "./run-progress-stepper"
+import { TimelineItem } from "./run-timeline-item"
+import { InfoBanner } from "./run-info-banner"
+import { FinishRunForm } from "./finish-run-form"
+import { CompleteRunForm } from "./complete-run-form"
+import { InlineTaskCard } from "./inline-task-card"
 
 /**
  * The work surface for a single production run: status, progress stepper,
@@ -380,7 +120,6 @@ export const ProductionRunCard = ({
   )
 
   const targetDateInfo = getTargetDateStatus((design as any)?.target_completion_date)
-  const guidance = getStageGuidance(run, isSample)
 
   // ── Actions ──
 
@@ -422,16 +161,21 @@ export const ProductionRunCard = ({
 
   const handleCompleteClick = () => setShowCompleteForm(true)
 
-  // Derive the primary action
-  const primaryAction = canAccept
-    ? { label: "Accept Run", onClick: handleAccept, loading: accept.isPending }
-    : canStart
-    ? { label: "Start Working", onClick: handleStart, loading: start.isPending }
-    : canFinish
-    ? { label: "Mark Finished", onClick: handleFinishClick, loading: finish.isPending }
-    : canComplete
-    ? { label: "Complete Run", onClick: handleCompleteClick, loading: complete.isPending }
-    : null
+  /**
+   * #2018 — `RunNextAction` decides WHICH action to offer (one derivation, unit
+   * tested); the card still owns how to run it. This map is the whole seam.
+   *
+   * The old code derived the button here from `can*` and the guidance banner
+   * separately from a stage-guidance helper — the same state, twice.
+   */
+  const runAction = (key: RunActionKey) => {
+    if (key === "accept") return handleAccept()
+    if (key === "start") return handleStart()
+    if (key === "finish") return handleFinishClick()
+    return handleCompleteClick()
+  }
+  const actionPending =
+    accept.isPending || start.isPending || finish.isPending || complete.isPending
 
   // ── Timeline entries ──
 
@@ -514,18 +258,17 @@ export const ProductionRunCard = ({
             </div>
           )}
         </div>
-        <div className="flex items-center gap-x-2">
-          {primaryAction && (
-            <Button
-              size="small"
-              isLoading={primaryAction.loading}
-              onClick={primaryAction.onClick}
-            >
-              {primaryAction.label}
-            </Button>
-          )}
-        </div>
       </div>
+
+      {/* #2018 — the action first. Above the reassignment notice, the stepper
+          and every section: this is what the partner came here to do. */}
+      <RunNextAction
+        run={run}
+        isSample={isSample}
+        actionable={actionable}
+        isPending={actionPending}
+        onAction={runAction}
+      />
 
       {/* #1228 — how this run reached the partner. Sits ABOVE the stepper: a
           re-sent run and a fresh one are otherwise identical on screen, and
@@ -541,11 +284,6 @@ export const ProductionRunCard = ({
 
       {/* Progress stepper */}
       <ProgressStepper run={run} />
-
-      {/* Contextual guidance — shown between stepper and forms so partner reads it before acting */}
-      {guidance && !isCancelled && (
-        <InfoBanner title={guidance.title} description={guidance.description} />
-      )}
 
       {/* Sample run emphasis — shown from accept onwards (not just after start) */}
       {isSample && !isCancelled && !isCompleted && run.accepted_at && (
@@ -777,922 +515,5 @@ export const ProductionRunCard = ({
         </div>
       )}
     </Container>
-  )
-}
-
-// ── Finish Run Form ─────────────────────────────────────────────────
-
-const FinishRunForm = ({
-  open,
-  onOpenChange,
-  pendingTasks,
-  finishNotes,
-  setFinishNotes,
-  onConfirm,
-  isLoading,
-  isSample,
-  consumptionCount,
-}: {
-  open: boolean
-  onOpenChange: (open: boolean) => void
-  pendingTasks: any[]
-  finishNotes: string
-  setFinishNotes: (v: string) => void
-  onConfirm: () => void
-  isLoading: boolean
-  isSample: boolean
-  consumptionCount: number
-}) => {
-  const [acknowledgedPending, setAcknowledgedPending] = useState(false)
-  const hasPending = pendingTasks.length > 0
-  const canConfirm = !hasPending || acknowledgedPending
-
-  return (
-    <Drawer open={open} onOpenChange={onOpenChange}>
-      <Drawer.Content>
-        <Drawer.Header>
-          <Drawer.Title>Mark as Finished</Drawer.Title>
-          <Drawer.Description>
-            The design will move to Technical Review for admin to inspect.
-          </Drawer.Description>
-        </Drawer.Header>
-        <Drawer.Body className="flex flex-col gap-y-4 overflow-y-auto">
-          {/* Sample run warning if no materials logged */}
-          {isSample && consumptionCount === 0 && (
-            <div className="flex items-start gap-x-3 rounded-xl border border-ui-border-base bg-ui-bg-subtle px-4 py-3">
-              <ExclamationCircle className="mt-0.5 shrink-0 text-ui-tag-orange-icon" />
-              <div className="flex flex-col gap-y-0.5">
-                <Text size="small" weight="plus">No materials logged yet</Text>
-                <Text size="xsmall" className="text-ui-fg-subtle">
-                  For sample runs, material usage data is needed for cost estimation. Consider logging materials before finishing.
-                </Text>
-              </div>
-            </div>
-          )}
-
-          {hasPending && (
-            <div className="rounded-xl border border-ui-border-base bg-ui-bg-subtle px-4 py-3">
-              <Text size="small" weight="plus" className="text-ui-fg-subtle mb-1">
-                {pendingTasks.length} task(s) still pending
-              </Text>
-              <div className="flex flex-col gap-y-0.5 mb-3">
-                {pendingTasks.map((t: any) => (
-                  <Text key={t.id} size="xsmall" className="text-ui-fg-muted">
-                    &bull; {t.title || t.id}
-                  </Text>
-                ))}
-              </div>
-              <label className="flex items-center gap-2 cursor-pointer">
-                <Checkbox
-                  checked={acknowledgedPending}
-                  onCheckedChange={(checked) => setAcknowledgedPending(!!checked)}
-                />
-                <Text size="xsmall">
-                  I confirm these tasks are completed or not needed
-                </Text>
-              </label>
-            </div>
-          )}
-
-          <div>
-            <Text size="xsmall" weight="plus" className="text-ui-fg-subtle mb-1">
-              Notes for reviewer (optional)
-            </Text>
-            <Textarea
-              placeholder="Any notes about the finished work, issues encountered, etc."
-              value={finishNotes}
-              onChange={(e) => setFinishNotes(e.target.value)}
-              rows={3}
-            />
-          </div>
-        </Drawer.Body>
-        <Drawer.Footer>
-          <Button variant="secondary" size="small" onClick={() => onOpenChange(false)}>
-            Cancel
-          </Button>
-          <Button
-            size="small"
-            isLoading={isLoading}
-            onClick={onConfirm}
-            disabled={!canConfirm}
-          >
-            Confirm & Mark Finished
-          </Button>
-        </Drawer.Footer>
-      </Drawer.Content>
-    </Drawer>
-  )
-}
-
-// ── Complete Run Form ───────────────────────────────────────────────
-
-const UNIT_OPTIONS = [
-  { value: "Meter", label: "Meter" },
-  { value: "Yard", label: "Yard" },
-  { value: "Kilogram", label: "Kilogram" },
-  { value: "Gram", label: "Gram" },
-  { value: "Piece", label: "Piece" },
-  { value: "Roll", label: "Roll" },
-  { value: "Other", label: "Other" },
-]
-
-type ConsumptionEntry = {
-  inventory_item_id: string
-  quantity: string
-  unit_cost: string
-  unit_of_measure: string
-  notes: string
-}
-
-const REJECTION_REASONS = [
-  { value: "stitching_defect", label: "Stitching defect" },
-  { value: "fabric_flaw", label: "Fabric flaw" },
-  { value: "color_mismatch", label: "Color mismatch" },
-  { value: "sizing_error", label: "Sizing error" },
-  { value: "print_defect", label: "Print defect" },
-  { value: "material_damage", label: "Material damage" },
-  { value: "quality_below_standard", label: "Quality below standard" },
-  { value: "other", label: "Other" },
-]
-
-const CompleteRunForm = ({
-  run,
-  design,
-  onComplete,
-  open,
-  onOpenChange,
-  isLoading,
-  isSample,
-  existingConsumptionCount,
-}: {
-  run: any
-  design: PartnerDesign
-  onComplete: (body: any) => Promise<void>
-  open: boolean
-  onOpenChange: (open: boolean) => void
-  isLoading: boolean
-  isSample: boolean
-  existingConsumptionCount: number
-}) => {
-  // Only what THIS run was assigned. Falls back to the design's full bill of
-  // materials when the run carries no allocation — which is every run made
-  // before assignments could carry materials.
-  const { options: inventoryItems, constrained: materialsConstrained } =
-    resolveRunMaterialOptions(run, design)
-  const tasks = run.tasks || []
-  const pendingTasks = tasks.filter(
-    (t: any) => t.status !== "completed" && t.status !== "cancelled"
-  )
-  const prompt = usePrompt()
-  const runQuantity = run.quantity || 1
-
-  // ── Step 1: Output ──
-  const [producedQty, setProducedQty] = useState(String(runQuantity))
-  const [rejectedQty, setRejectedQty] = useState("")
-  const [rejectionReason, setRejectionReason] = useState("")
-  const [rejectionNotes, setRejectionNotes] = useState("")
-  /**
-   * #1271 — what happened to units that were neither made nor rejected.
-   * The backend allows a genuine shortfall, but only with `allow_shortfall`
-   * AND a written explanation; without a control for it the only way past the
-   * gate was to inflate `produced_quantity`, which is what the gate exists to
-   * prevent.
-   */
-  const [shortfallReason, setShortfallReason] = useState("")
-
-  /**
-   * ── Step 2: Cost ──
-   *
-   * 🔴 Starts UNSET, not "total". It used to default to "total" with that
-   * button highlighted, so a partner typing their PER-PIECE rate produced a run
-   * labelled `cost_type: "total"` — and the payout billed it once: ₹850 for
-   * nine garments, with nothing erroring and the number looking plausible.
-   * The two readings differ by a factor of the run quantity, so there is no
-   * safe default; the backend now refuses a cost with no type. (#1554)
-   */
-  const [costType, setCostType] = useState<"per_unit" | "total" | null>(null)
-  const [partnerEstimate, setPartnerEstimate] = useState("")
-
-  // ── Step 3: Additional materials ──
-  const [showMaterialForm, setShowMaterialForm] = useState(false)
-  const [consumptions, setConsumptions] = useState<ConsumptionEntry[]>([])
-
-  const addConsumptionItem = (itemId: string) => {
-    if (consumptions.some((c) => c.inventory_item_id === itemId)) return
-    setConsumptions((prev) => [...prev, {
-      inventory_item_id: itemId,
-      quantity: "",
-      unit_cost: "",
-      unit_of_measure: "Meter",
-      notes: "",
-    }])
-  }
-
-  const removeConsumptionItem = (idx: number) => {
-    setConsumptions((prev) => prev.filter((_, i) => i !== idx))
-  }
-
-  // Items not yet added to the form
-  const availableItems = inventoryItems.filter(
-    (item: any) => !consumptions.some((c) => c.inventory_item_id === item.id)
-  )
-
-  // ── Step 4: Notes ──
-  const [completionNotes, setCompletionNotes] = useState("")
-
-  // Derived values
-  const produced = parseFloat(producedQty) || 0
-  const rejected = parseFloat(rejectedQty) || 0
-  const yieldPct = runQuantity > 0 ? Math.round((produced / runQuantity) * 100) : 0
-  // Rejects are output too — output that failed. Anything left over is
-  // unaccounted for, and the completion gate refuses it unless it is declared.
-  const outputPlan = planCompletionOutput({
-    ordered: runQuantity,
-    produced,
-    rejected,
-    shortfallReason,
-  })
-  const unaccounted = outputPlan.unaccounted
-  const costValue = parseFloat(partnerEstimate) || 0
-  /**
-   * 🔴 The payout multiplier is the ORDERED quantity, not the produced one —
-   * see `runPayableAmount`: correcting a partner's output figure deliberately
-   * does not move the money. This preview used `produced`, so on any run where
-   * the two differed it showed a total the payment would never match.
-   */
-  const costUnits = runQuantity > 0 ? runQuantity : 1
-  const totalCost = costType === "per_unit" ? Math.round(costValue * costUnits * 100) / 100 : costValue
-  const perUnitCost = costType === "total" && costUnits > 0 ? Math.round(costValue / costUnits * 100) / 100 : costValue
-
-  const updateConsumption = (idx: number, field: keyof ConsumptionEntry, value: string) => {
-    setConsumptions((prev) =>
-      prev.map((c, i) => (i === idx ? { ...c, [field]: value } : c))
-    )
-  }
-
-  const buildBody = () => {
-    const validConsumptions = consumptions
-      .filter((c) => c.quantity && parseFloat(c.quantity) > 0)
-      .map((c) => ({
-        inventory_item_id: c.inventory_item_id,
-        quantity: parseFloat(c.quantity),
-        unit_cost: c.unit_cost ? parseFloat(c.unit_cost) : undefined,
-        unit_of_measure: c.unit_of_measure,
-        notes: c.notes || undefined,
-      }))
-
-    // A cost with no type is refused by the backend (#1554) — say so here
-    // rather than letting the partner submit and read a 400. The two readings
-    // differ by the run quantity, so neither can be assumed.
-    if (costValue > 0 && !costType) {
-      toast.error("Choose whether the cost is per piece or for the whole run")
-      return
-    }
-
-    const body: any = {
-      produced_quantity: produced,
-    }
-    if (rejected > 0) {
-      body.rejected_quantity = rejected
-      if (rejectionReason) body.rejection_reason = rejectionReason
-      if (rejectionNotes.trim()) body.rejection_notes = rejectionNotes.trim()
-    }
-    if (costValue > 0) {
-      body.partner_cost_estimate = costValue
-      body.cost_type = costType
-    }
-    if (validConsumptions.length > 0) body.consumptions = validConsumptions
-
-    // The shortfall explanation rides in `notes` — that is the field the gate
-    // reads, and it keeps the reason with the completion rather than in a
-    // side channel nobody looks at.
-    const noteParts = [outputPlan.noteLine || "", completionNotes.trim()].filter(
-      Boolean
-    )
-
-    if (noteParts.length) body.notes = noteParts.join("\n")
-    if (outputPlan.allowShortfall) body.allow_shortfall = true
-
-    return body
-  }
-
-  const handleSubmit = async () => {
-    // Mirror the backend gate so the partner gets a control, not a 400 telling
-    // them to do something the form could not express (#1271).
-    if (outputPlan.needsReason) {
-      toast.error(
-        `${unaccounted} piece${unaccounted !== 1 ? "s" : ""} unaccounted for`,
-        {
-          description:
-            "Say what happened to them — record them as rejected, or write the reason in the shortfall box. Don't raise the produced count to cover them.",
-        }
-      )
-      return
-    }
-
-    const body = buildBody()
-
-    // Warn if no cost data
-    if (!costValue && !isSample) {
-      const confirmed = await prompt({
-        title: "Complete without a cost estimate?",
-        description: "No production cost entered. This helps with pricing and margin tracking.",
-        confirmText: "Complete Anyway",
-        cancelText: "Go Back",
-      })
-      if (!confirmed) return
-    }
-
-    // Sample runs: stricter — warn if no cost AND no materials
-    if (isSample && !costValue && existingConsumptionCount === 0 && !body.consumptions?.length) {
-      const confirmed = await prompt({
-        title: "Complete sample without cost or material data?",
-        description: "Sample runs need this data for cost estimation. Are you sure?",
-        confirmText: "Complete Anyway",
-        cancelText: "Go Back",
-      })
-      if (!confirmed) return
-    }
-
-    await onComplete(body)
-  }
-
-  return (
-    <FocusModal open={open} onOpenChange={onOpenChange}>
-      <FocusModal.Content>
-        <FocusModal.Header>
-          <FocusModal.Title>Complete Production Run</FocusModal.Title>
-          <FocusModal.Description>
-            {runQuantity} piece{runQuantity !== 1 ? "s" : ""} ordered
-            {isSample ? " · Sample run" : ""}
-          </FocusModal.Description>
-        </FocusModal.Header>
-        <FocusModal.Body className="flex flex-col items-center overflow-y-auto py-6">
-          <div className="flex w-full max-w-2xl flex-col gap-y-4 px-4">
-      {pendingTasks.length > 0 && (
-        <div className="rounded-xl border border-ui-border-base bg-ui-bg-base px-4 py-3 mb-4">
-          <Text size="small" weight="plus" className="text-ui-fg-subtle mb-1">
-            {pendingTasks.length} pending task(s) will be marked as done
-          </Text>
-          <div className="flex flex-col gap-y-0.5">
-            {pendingTasks.map((t: any) => (
-              <Text key={t.id} size="xsmall" className="text-ui-fg-muted">
-                &bull; {t.title || t.id}
-              </Text>
-            ))}
-          </div>
-        </div>
-      )}
-
-      {/* ── Step 1: Output ── */}
-      <div className="rounded-xl border border-ui-border-base bg-ui-bg-base px-4 py-4 mb-4">
-        <Text size="small" weight="plus" className="mb-3">Output</Text>
-        <div className="grid grid-cols-2 gap-4">
-          <div>
-            <Text size="xsmall" className="text-ui-fg-subtle mb-1">
-              Good pieces produced
-            </Text>
-            <Input
-              type="number"
-              min="0"
-              max={runQuantity}
-              step="1"
-              value={producedQty}
-              onChange={(e) => setProducedQty(e.target.value)}
-            />
-          </div>
-          <div>
-            <Text size="xsmall" className="text-ui-fg-subtle mb-1">
-              Rejected
-            </Text>
-            <Input
-              type="number"
-              min="0"
-              step="1"
-              value={rejectedQty}
-              onChange={(e) => setRejectedQty(e.target.value)}
-            />
-          </div>
-        </div>
-
-        {/* Yield indicator */}
-        {produced > 0 && (
-          <div className="flex items-center gap-3 mt-3 pt-3 border-t border-ui-border-base">
-            <StatusBadge color={yieldPct >= 90 ? "green" : yieldPct >= 70 ? "orange" : "red"}>
-              {yieldPct}% yield
-            </StatusBadge>
-            <Text size="xsmall" className="text-ui-fg-muted">
-              {produced} of {runQuantity} pieces
-              {rejected > 0 ? ` · ${rejected} rejected` : ""}
-              {unaccounted > 0 ? ` · ${unaccounted} unaccounted for` : ""}
-            </Text>
-          </div>
-        )}
-
-        {/*
-          #1271 — the missing units. The form used to fill the rejected field
-          with the remainder automatically, which recorded units that were
-          never made as units that FAILED, and left a partner whose shortfall
-          was something else (fabric ran out, order cut short) with no way
-          through the completion gate except inflating the produced count.
-        */}
-        {unaccounted > 0 && (
-          <div className="mt-3 pt-3 border-t border-ui-border-base">
-            <div className="flex items-center justify-between gap-3 mb-2">
-              <Text size="xsmall" weight="plus" className="text-ui-tag-orange-text">
-                {unaccounted} piece{unaccounted !== 1 ? "s" : ""} not accounted for
-              </Text>
-              <Button
-                type="button"
-                size="small"
-                variant="secondary"
-                onClick={() => {
-                  setRejectedQty(String(rejected + unaccounted))
-                  setShortfallReason("")
-                }}
-              >
-                They were rejected
-              </Button>
-            </div>
-            <Text size="xsmall" className="text-ui-fg-subtle mb-1">
-              If they failed quality, record them as rejected. Otherwise say what
-              happened — a recorded shortfall is fine, a wrong produced count is not.
-            </Text>
-            <Textarea
-              rows={2}
-              placeholder="e.g. fabric ran out after 8 pieces — remaining 2 not cut"
-              value={shortfallReason}
-              onChange={(e) => setShortfallReason(e.target.value)}
-            />
-          </div>
-        )}
-
-        {/* Rejection details */}
-        {rejected > 0 && (
-          <div className="mt-3 pt-3 border-t border-ui-border-base">
-            <div className="grid grid-cols-2 gap-3">
-              <div>
-                <Text size="xsmall" className="text-ui-fg-subtle mb-1">Reason</Text>
-                <Select value={rejectionReason} onValueChange={setRejectionReason}>
-                  <Select.Trigger>
-                    <Select.Value placeholder="Select reason" />
-                  </Select.Trigger>
-                  <Select.Content>
-                    {REJECTION_REASONS.map((r) => (
-                      <Select.Item key={r.value} value={r.value}>{r.label}</Select.Item>
-                    ))}
-                  </Select.Content>
-                </Select>
-              </div>
-              <div>
-                <Text size="xsmall" className="text-ui-fg-subtle mb-1">Details (optional)</Text>
-                <Input
-                  placeholder="e.g. thread pull on collar"
-                  value={rejectionNotes}
-                  onChange={(e) => setRejectionNotes(e.target.value)}
-                />
-              </div>
-            </div>
-          </div>
-        )}
-      </div>
-
-      {/* ── Step 2: Cost ── */}
-      <div className="rounded-xl border border-ui-border-base bg-ui-bg-base px-4 py-4 mb-4">
-        <Text size="small" weight="plus" className="mb-1">
-          Your production cost
-          {isSample && <span className="text-ui-fg-error ml-1">*</span>}
-        </Text>
-        <Text size="xsmall" className="text-ui-fg-subtle mb-3">
-          {isSample
-            ? "Your charge for producing this sample. This feeds directly into pricing."
-            : "Your charge for this production run (labor, overheads, margin)."
-          }
-        </Text>
-
-        <div className="flex items-end gap-3">
-          <div className="flex-1 max-w-[200px]">
-            <Input
-              type="number"
-              step="0.01"
-              min="0"
-              placeholder="0.00"
-              value={partnerEstimate}
-              onChange={(e) => setPartnerEstimate(e.target.value)}
-            />
-          </div>
-          <div className="flex rounded-lg border border-ui-border-base overflow-hidden">
-            <button
-              type="button"
-              className={clx(
-                "px-3 py-2 text-xs transition-colors",
-                costType === "per_unit"
-                  ? "bg-ui-bg-interactive text-ui-fg-on-color"
-                  : "bg-ui-bg-base text-ui-fg-subtle hover:bg-ui-bg-base-hover"
-              )}
-              onClick={() => setCostType("per_unit")}
-            >
-              Per piece
-            </button>
-            <button
-              type="button"
-              className={clx(
-                "px-3 py-2 text-xs transition-colors border-l border-ui-border-base",
-                costType === "total"
-                  ? "bg-ui-bg-interactive text-ui-fg-on-color"
-                  : "bg-ui-bg-base text-ui-fg-subtle hover:bg-ui-bg-base-hover"
-              )}
-              onClick={() => setCostType("total")}
-            >
-              Total
-            </button>
-          </div>
-        </div>
-
-        {/* What this run will actually pay.
-            🔴 Stated against the ORDERED quantity, because that is the
-            multiplier the payout uses (`runPayableAmount`). The old line used
-            the produced figure, so on any run where the two differed it
-            previewed a total the payment would never match. */}
-        {costValue > 0 && costType && (
-          <div className="flex items-center gap-3 mt-3 pt-3 border-t border-ui-border-base">
-            <Text size="xsmall" className="text-ui-fg-muted">
-              {(() => {
-                const cur = ((design as any)?.cost_currency || "").toUpperCase()
-                const prefix = cur ? `${cur} ` : ""
-                return costType === "per_unit"
-                  ? `${prefix}${costValue} × ${costUnits} ordered = ${prefix}${totalCost} total`
-                  : `${prefix}${totalCost} total · ${prefix}${perUnitCost} per piece of ${costUnits}`
-              })()}
-            </Text>
-          </div>
-        )}
-
-        {/* The unchosen state is a real state, and silence about it is how the
-            per-piece rate got billed once. */}
-        {costValue > 0 && !costType && (
-          <div className="flex items-center gap-3 mt-3 pt-3 border-t border-ui-border-base">
-            <Text size="xsmall" className="text-ui-fg-error">
-              Is that per piece, or for all {costUnits}? The two are paid very
-              differently.
-            </Text>
-          </div>
-        )}
-      </div>
-
-      {/* ── Step 3: Materials ── */}
-      <div className="rounded-xl border border-ui-border-base bg-ui-bg-base px-4 py-4 mb-4">
-        <div className="flex items-center justify-between mb-1">
-          <Text size="small" weight="plus">
-            Materials
-            {isSample && <span className="text-ui-fg-error ml-1">*</span>}
-          </Text>
-          {existingConsumptionCount > 0 && (
-            <Badge size="2xsmall" color="green">
-              {existingConsumptionCount} already logged
-            </Badge>
-          )}
-        </div>
-
-        {materialsConstrained && (
-          <Text size="xsmall" className="text-ui-fg-subtle mb-2">
-            This run was assigned specific materials
-            {inventoryItems.some((i) => i.planned_quantity != null)
-              ? ` — ${inventoryItems
-                  .filter((i) => i.planned_quantity != null)
-                  .map((i) => `${i.title || i.sku || i.id}: ${i.planned_quantity}`)
-                  .join(", ")}`
-              : ""}
-            . Only these can be logged against it.
-          </Text>
-        )}
-
-        {existingConsumptionCount > 0 ? (
-          <Text size="xsmall" className="text-ui-fg-subtle mb-3">
-            {existingConsumptionCount} material entry(ies) were logged during production.
-            {inventoryItems.length > 0 ? " Add more below if needed." : ""}
-          </Text>
-        ) : (
-          <Text size="xsmall" className="text-ui-fg-subtle mb-3">
-            {isSample
-              ? "No materials logged yet. Record what was consumed — this data is critical for cost estimation."
-              : "Log materials consumed during this run (optional)."
-            }
-          </Text>
-        )}
-
-        {/* Added items */}
-        {consumptions.length > 0 && (
-          <div className="flex flex-col gap-3 mt-3">
-            {consumptions.map((entry, idx) => {
-              const item = inventoryItems.find((i: any) => i.id === entry.inventory_item_id)
-              const label = item?.title || item?.sku || entry.inventory_item_id
-              return (
-                <div key={entry.inventory_item_id} className="rounded-lg border border-ui-border-base p-3">
-                  <div className="flex items-center justify-between mb-2">
-                    <Text size="xsmall" weight="plus">{label}</Text>
-                    <Button variant="transparent" size="small" onClick={() => removeConsumptionItem(idx)}>
-                      Remove
-                    </Button>
-                  </div>
-                  <div className="grid grid-cols-2 gap-2 md:grid-cols-4">
-                    <div>
-                      <Text size="xsmall" className="text-ui-fg-subtle mb-1">Qty</Text>
-                      <Input
-                        type="number"
-                        step="0.01"
-                        min="0"
-                        placeholder="0"
-                        value={entry.quantity}
-                        onChange={(e) => updateConsumption(idx, "quantity", e.target.value)}
-                      />
-                    </div>
-                    <div>
-                      <Text size="xsmall" className="text-ui-fg-subtle mb-1">Cost/unit</Text>
-                      <Input
-                        type="number"
-                        step="0.01"
-                        min="0"
-                        placeholder="Optional"
-                        value={entry.unit_cost}
-                        onChange={(e) => updateConsumption(idx, "unit_cost", e.target.value)}
-                      />
-                    </div>
-                    <div>
-                      <Text size="xsmall" className="text-ui-fg-subtle mb-1">Unit</Text>
-                      <Select value={entry.unit_of_measure} onValueChange={(v) => updateConsumption(idx, "unit_of_measure", v)}>
-                        <Select.Trigger><Select.Value /></Select.Trigger>
-                        <Select.Content>
-                          {UNIT_OPTIONS.map((o) => (
-                            <Select.Item key={o.value} value={o.value}>{o.label}</Select.Item>
-                          ))}
-                        </Select.Content>
-                      </Select>
-                    </div>
-                    <div>
-                      <Text size="xsmall" className="text-ui-fg-subtle mb-1">Notes</Text>
-                      <Input
-                        placeholder="Optional"
-                        value={entry.notes}
-                        onChange={(e) => updateConsumption(idx, "notes", e.target.value)}
-                      />
-                    </div>
-                  </div>
-                </div>
-              )
-            })}
-          </div>
-        )}
-
-        {/* Add item selector */}
-        {availableItems.length > 0 && (
-          <div className="mt-3">
-            <Select onValueChange={(v) => { addConsumptionItem(v); setShowMaterialForm(true) }}>
-              <Select.Trigger>
-                <Select.Value placeholder="Add an inventory item..." />
-              </Select.Trigger>
-              <Select.Content>
-                {availableItems.map((item: any) => (
-                  <Select.Item key={item.id} value={item.id}>
-                    {item.title || item.sku || item.id}
-                  </Select.Item>
-                ))}
-              </Select.Content>
-            </Select>
-          </div>
-        )}
-      </div>
-
-      {/* ── Step 4: Notes ── */}
-      <div className="mb-4">
-        <Text size="xsmall" weight="plus" className="text-ui-fg-subtle mb-1">
-          Completion notes (optional)
-        </Text>
-        <Textarea
-          placeholder="Quality observations, issues, or feedback for the admin team"
-          value={completionNotes}
-          onChange={(e) => setCompletionNotes(e.target.value)}
-          rows={3}
-        />
-      </div>
-
-          </div>
-        </FocusModal.Body>
-        <FocusModal.Footer>
-          <Button variant="secondary" size="small" onClick={() => onOpenChange(false)} disabled={isLoading}>
-            Cancel
-          </Button>
-          <Button size="small" onClick={handleSubmit} isLoading={isLoading}>
-            Confirm & Complete
-          </Button>
-        </FocusModal.Footer>
-      </FocusModal.Content>
-    </FocusModal>
-  )
-}
-
-// ── Inline Task Card ────────────────────────────────────────────────
-
-const InlineTaskCard = ({ task, linkBase }: { task: any; linkBase?: string }) => {
-  const taskId = String(task.id)
-  const status = String(task.status || "pending")
-  const canAccept = status === "pending" || status === "assigned"
-  const canFinish = status === "accepted" || status === "in_progress"
-  const isCompleted = status === "completed"
-  const subtasks = task.subtasks || []
-
-  const [showCostInput, setShowCostInput] = useState(false)
-  const [actualCost, setActualCost] = useState("")
-
-  const acceptTask = useAcceptPartnerAssignedTask(taskId)
-  const finishTask = useFinishPartnerAssignedTask(taskId)
-
-  const estimatedCost = task.estimated_cost ? Number(task.estimated_cost) : null
-  const completedCost = task.actual_cost ? Number(task.actual_cost) : null
-
-  const handleAccept = async () => {
-    try {
-      await acceptTask.mutateAsync()
-      toast.success(`Task "${task.title}" accepted`)
-    } catch (e) {
-      toast.error(extractErrorMessage(e))
-    }
-  }
-
-  const handleFinishClick = () => {
-    // If task has an estimated cost, show cost input before finishing
-    if (estimatedCost || estimatedCost === 0) {
-      setShowCostInput(true)
-      setActualCost(estimatedCost ? String(estimatedCost) : "")
-    } else {
-      setShowCostInput(true)
-      setActualCost("")
-    }
-  }
-
-  const handleFinishConfirm = async () => {
-    try {
-      const cost = actualCost ? parseFloat(actualCost) : undefined
-      await finishTask.mutateAsync(
-        cost && cost > 0 ? { actual_cost: cost } : undefined
-      )
-      toast.success(`Task "${task.title}" finished`)
-      setShowCostInput(false)
-    } catch (e) {
-      toast.error(extractErrorMessage(e))
-    }
-  }
-
-  const handleFinishSkip = async () => {
-    try {
-      await finishTask.mutateAsync()
-      toast.success(`Task "${task.title}" finished`)
-      setShowCostInput(false)
-    } catch (e) {
-      toast.error(extractErrorMessage(e))
-    }
-  }
-
-  return (
-    <div className="rounded-lg border border-ui-border-base p-3">
-      <div className="flex items-start justify-between gap-x-3">
-        <div className="min-w-0 flex-1">
-          <div className="flex items-center gap-x-2">
-            {linkBase ? (
-              <Link
-                to={`${linkBase}/${taskId}`}
-                className="truncate text-ui-fg-interactive hover:text-ui-fg-interactive-hover"
-              >
-                <Text size="small" weight="plus" className="truncate">
-                  {String(task.title || task.id)}
-                </Text>
-              </Link>
-            ) : (
-              <Text size="small" weight="plus" className="truncate">
-                {String(task.title || task.id)}
-              </Text>
-            )}
-            <Badge size="2xsmall" color={getStatusBadgeColor(status)}>
-              {status.replace(/_/g, " ")}
-            </Badge>
-          </div>
-          {task.description && (
-            <Text size="xsmall" className="text-ui-fg-subtle mt-1">
-              {String(task.description)}
-            </Text>
-          )}
-          {/* Show cost info */}
-          {(estimatedCost || completedCost) && (
-            <div className="flex items-center gap-2 mt-1">
-              {estimatedCost != null && (
-                <Text size="xsmall" className="text-ui-fg-muted">
-                  Est: {estimatedCost}
-                </Text>
-              )}
-              {completedCost != null && (
-                <Text size="xsmall" weight="plus">
-                  Actual: {completedCost}
-                </Text>
-              )}
-            </div>
-          )}
-        </div>
-        <div className="flex shrink-0 items-center gap-x-2">
-          {canAccept && (
-            <Button size="small" variant="secondary" isLoading={acceptTask.isPending} onClick={handleAccept}>
-              Accept
-            </Button>
-          )}
-          {canFinish && !showCostInput && (
-            <Button size="small" isLoading={finishTask.isPending} onClick={handleFinishClick}>
-              Finish
-            </Button>
-          )}
-          {isCompleted && <Checkbox checked disabled className="mt-0.5" />}
-        </div>
-      </div>
-
-      {/* Cost input when finishing */}
-      {showCostInput && canFinish && (
-        <div className="mt-3 pt-3 border-t border-ui-border-base">
-          <Text size="xsmall" weight="plus" className="text-ui-fg-subtle mb-1">
-            Your cost for this work (optional)
-          </Text>
-          {estimatedCost != null && (
-            <Text size="xsmall" className="text-ui-fg-muted mb-2">
-              Estimated: {estimatedCost}
-            </Text>
-          )}
-          <div className="flex items-end gap-2">
-            <div className="max-w-[160px]">
-              <Input
-                type="number"
-                step="0.01"
-                min="0"
-                placeholder={estimatedCost ? String(estimatedCost) : "0.00"}
-                value={actualCost}
-                onChange={(e) => setActualCost(e.target.value)}
-              />
-            </div>
-            <Button size="small" variant="secondary" onClick={handleFinishSkip} isLoading={finishTask.isPending}>
-              Skip
-            </Button>
-            <Button size="small" onClick={handleFinishConfirm} isLoading={finishTask.isPending}>
-              Finish
-            </Button>
-          </div>
-        </div>
-      )}
-
-      {subtasks.length > 0 && (
-        <div className="mt-3 border-t pt-3">
-          <Text size="xsmall" weight="plus" className="text-ui-fg-subtle mb-2">
-            Subtasks ({subtasks.filter((s: any) => s.status === "completed").length}/{subtasks.length})
-          </Text>
-          <div className="flex flex-col gap-y-2">
-            {subtasks.map((sub: any) => (
-              <InlineSubtaskRow key={String(sub.id)} taskId={taskId} subtask={sub} />
-            ))}
-          </div>
-        </div>
-      )}
-    </div>
-  )
-}
-
-// ── Inline Subtask Row ──────────────────────────────────────────────
-
-const InlineSubtaskRow = ({ taskId, subtask }: { taskId: string; subtask: any }) => {
-  const subtaskId = String(subtask.id)
-  const isCompleted = subtask.status === "completed"
-  const complete = useCompletePartnerAssignedTaskSubtask(taskId, subtaskId)
-
-  const handleComplete = async () => {
-    try {
-      await complete.mutateAsync()
-      toast.success(`Subtask "${subtask.title}" completed`)
-    } catch (e) {
-      toast.error(extractErrorMessage(e))
-    }
-  }
-
-  return (
-    <div className="flex items-center justify-between gap-x-3 rounded border border-ui-border-base px-3 py-2">
-      <div className="flex items-center gap-x-2 min-w-0">
-        <Checkbox
-          checked={isCompleted}
-          disabled={isCompleted || complete.isPending}
-          onCheckedChange={() => {
-            if (!isCompleted) handleComplete()
-          }}
-        />
-        <Text size="xsmall" className={clx({ "line-through text-ui-fg-muted": isCompleted })}>
-          {String(subtask.title || subtask.id)}
-        </Text>
-      </div>
-      <Badge size="2xsmall" color={getStatusBadgeColor(String(subtask.status))}>
-        {String(subtask.status).replace(/_/g, " ")}
-      </Badge>
-    </div>
   )
 }
